requirements.txt
asciidoc~=10.2.1
GitPython~=3.1.43
PySignal>=1.1.1
PyQt6~=6.7.1
PyQt6-WebEngine~=6.7.0
jaro-winkler~=2.0.3
